+++ sciris/sc_version.py
@@ -4,6 +4,6 @@ Version and license information.
 
 __all__ = ['__version__', '__versiondate__', '__license__']
 
-__version__      = '1.1.2'
+__version__      = '1.2.0'
 __versiondate__  = '2021-07-05'
 __license__      = f'Sciris {__version__} ({__versiondate__}) – © 2021 by Sciris.org'